renaming some modules. changed version to 0.4.
