+++ gettsim/functions.py
@@ -2,27 +2,14 @@
 in the documentation.
 
 """
-from gettsim.benefits.arbeitsl_geld import *  # noqa: F401, F403
-from gettsim.benefits.arbeitsl_geld_2.arbeitsl_geld_2 import *  # noqa: F401, F403
-from gettsim.benefits.arbeitsl_geld_2.arbeitsl_geld_2_eink import *  # noqa: F401, F403
-from gettsim.benefits.arbeitsl_geld_2.eink_anr_frei import *  # noqa: F401, F403
-from gettsim.benefits.arbeitsl_geld_2.kost_unterk import *  # noqa: F401, F403
-from gettsim.benefits.benefit_checks.benefit_checks import *  # noqa: F401, F403
-from gettsim.benefits.benefit_checks.vermoegens_checks import *  # noqa: F401, F403
-from gettsim.benefits.elterngeld import *  # noqa: F401, F403
-from gettsim.benefits.kinderzuschlag.kinderzuschlag import *  # noqa: F401, F403
-from gettsim.benefits.kinderzuschlag.kinderzuschlag_eink import *  # noqa: F401, F403
-from gettsim.benefits.kinderzuschlag.kost_unterk import *  # noqa: F401, F403
-from gettsim.benefits.unterhalt import *  # noqa: F401, F403
-from gettsim.benefits.wohngeld import *  # noqa: F401, F403
 from gettsim.demographic_vars import *  # noqa: F401, F403
 from gettsim.renten_anspr import *  # noqa: F401, F403
-from gettsim.soz_vers.arbeitsl_v import *  # noqa: F401, F403
-from gettsim.soz_vers.beitr_bemess_grenzen import *  # noqa: F401, F403
-from gettsim.soz_vers.eink_grenzen import *  # noqa: F401, F403
-from gettsim.soz_vers.krankenv import *  # noqa: F401, F403
-from gettsim.soz_vers.pflegev import *  # noqa: F401, F403
-from gettsim.soz_vers.rentenv import *  # noqa: F401, F403
+from gettsim.social_insurance.arbeitsl_v import *  # noqa: F401, F403
+from gettsim.social_insurance.beitr_bemess_grenzen import *  # noqa: F401, F403
+from gettsim.social_insurance.eink_grenzen import *  # noqa: F401, F403
+from gettsim.social_insurance.krankenv import *  # noqa: F401, F403
+from gettsim.social_insurance.pflegev import *  # noqa: F401, F403
+from gettsim.social_insurance.rentenv import *  # noqa: F401, F403
 from gettsim.taxes.abgelt_st import *  # noqa: F401, F403
 from gettsim.taxes.eink_st import *  # noqa: F401, F403
 from gettsim.taxes.favorability_check import *  # noqa: F401, F403
@@ -32,3 +19,16 @@ from gettsim.taxes.zu_verst_eink.eink import *  # noqa: F401, F403
 from gettsim.taxes.zu_verst_eink.freibetraege import *  # noqa: F401, F403
 from gettsim.taxes.zu_verst_eink.vorsorge import *  # noqa: F401, F403
 from gettsim.taxes.zu_verst_eink.zu_verst_eink import *  # noqa: F401, F403
+from gettsim.transfers.arbeitsl_geld import *  # noqa: F401, F403
+from gettsim.transfers.arbeitsl_geld_2.arbeitsl_geld_2 import *  # noqa: F401, F403
+from gettsim.transfers.arbeitsl_geld_2.arbeitsl_geld_2_eink import *  # noqa: F401, F403
+from gettsim.transfers.arbeitsl_geld_2.eink_anr_frei import *  # noqa: F401, F403
+from gettsim.transfers.arbeitsl_geld_2.kost_unterk import *  # noqa: F401, F403
+from gettsim.transfers.benefit_checks.benefit_checks import *  # noqa: F401, F403
+from gettsim.transfers.benefit_checks.vermoegens_checks import *  # noqa: F401, F403
+from gettsim.transfers.elterngeld import *  # noqa: F401, F403
+from gettsim.transfers.kinderzuschlag.kinderzuschlag import *  # noqa: F401, F403
+from gettsim.transfers.kinderzuschlag.kinderzuschlag_eink import *  # noqa: F401, F403
+from gettsim.transfers.kinderzuschlag.kost_unterk import *  # noqa: F401, F403
+from gettsim.transfers.unterhalt import *  # noqa: F401, F403
+from gettsim.transfers.wohngeld import *  # noqa: F401, F403
