+++ gettsim/parameters/eink_st.yaml
