+++ gettsim/parameters/soli_st.yaml
