+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.4
+current_version = 0.4
 
 [bumpversion:file:setup.py]
